# perf: prefer parse int with + instead parseInt()