--- a/lib/textParsers.js
+++ b/lib/textParsers.js
@@ -18,12 +18,8 @@ function parseBoolArray (value) {
   return array.parse(value, parseBool)
 }
 
-function parseBaseTenInt (string) {
-  return parseInt(string, 10)
-}
-
 function parseIntegerArray (value) {
-  return array.parse(value, parseBaseTenInt)
+  return array.parse(value, Number)
 }
 
 function parseBigIntegerArray (value) {
@@ -120,7 +116,7 @@ const parseCircle = function (value) {
 }
 
 function parseInt4Range (raw) {
-  return range.parse(raw, parseBaseTenInt)
+  return range.parse(raw, Number)
 }
 
 function parseNumRange (raw) {
@@ -141,9 +137,9 @@ function parseTimestampTzRange (raw) {
 
 const init = function (register) {
   register(20, parseBigInteger) // int8
-  register(21, parseBaseTenInt) // int2
-  register(23, parseBaseTenInt) // int4
-  register(26, parseBaseTenInt) // oid
+  register(21, Number) // int2
+  register(23, Number) // int4
+  register(26, Number) // oid
   register(700, parseFloat) // float4/real
   register(701, parseFloat) // float8/double
   register(16, parseBool)