@charmander 2023-02-11
The input can’t be any of those values, so how about replacing `parseInteger` with `Number`?