@H4ad 2023-02-12
@charmander I think that could be possible but for the sake of compatibility with the old method, I think the performance will not be hurt to keep as it is with that `if` guard against `null`.