@charmander 2023-02-13
Why do we care about compatibility with these specific out-of-scope cases and not the rest of them? See also #122.