@H4ad 2023-02-13
@charmander In this case, make sense to me to just use `Number` instead, @bendrucker what do you think?